+++ YogaKit/Source/UIView+Yoga.h
@@ -10,8 +10,20 @@
 #import "YGLayout.h"
 #import <UIKit/UIKit.h>
 
+typedef void (^YGLayoutConfigurationBlock)(YGLayout *);
+
 @interface UIView (Yoga)
 
+/**
+ The YGLayout that is attached to this view. It is lazily created.
+ */
 @property (nonatomic, readonly, strong) YGLayout *yoga;
+
+/**
+ In ObjC land, every time you access `view.yoga.*` you are adding another `objc_msgSend`
+ to your code. If you plan on making multiple changes to YGLayout, it's more performant
+ to use this method, which uses a single objc_msgSend call.
+ */ 
+- (void)configureLayoutWithBlock:(YGLayoutConfigurationBlock)block NS_SWIFT_NAME(configureLayout(:));
 
 @end